Crate burn_ndarray

Burn ndarray backend.

Macros

iter_par

Macro for iterating in parallel.

iter_range_par

Macro for iterating over a range in parallel.

reshape

Reshapes an array into a tensor.

run_par

Macro for running a function in parallel.

to_typed_dims

Converts a slice of usize to a typed dimension.

Structs

NdArray

Tensor backend that uses the ndarray crate for executing tensor operations.

NdArrayTensor

Tensor primitive used by the ndarray backend.

Enums

NdArrayDevice

The device type for the ndarray backend.

Traits

FloatNdArrayElement

A float element for ndarray backend.